sha256:3410d0d07fa36193c319bb7849c975044d8d03ba168d5723a8b9800f975777e3

+++ goldrush/digger.go
@@ -51,7 +51,7 @@ func (d *Digger) run() {
 					fmt.Println(err)
 					return
 				}
-				d.measure.Add("depth_"+strconv.Itoa(depth)+"_sum", int32(len(treasures)))
+				d.measure.Add("depth_"+strconv.Itoa(depth)+"_sum", int64(len(treasures)))
 				for _, treasure := range treasures {
 					d.t.Cash(treasure)
 				}
